requirements.txt
protobuf==4.25.2
Requests==2.31.0
google-cloud-bigquery==3.17.1